Tighter RestGlobalContext yet more
Removes more getters in favor of pushing the HeadersAndContextCopyClient
into the global context. BaseRestHandler is now very very simple.
